+++ src/components/game/graphics/objects/Cursor.js
@@ -1,7 +1,7 @@
-import { DEBUG, MESSAGE_SIZE, MESSAGE_OFFSET, debugCell } from "../../../Debug";
-import { inBounds } from "../../../coreState/utils/Functions";
-import { outlineRect } from "../../Pipeline";
-import { FILLED_COLOR } from "../../Theme";
+import { DEBUG, MESSAGE_SIZE, MESSAGE_OFFSET, debugCell } from "../../Debug";
+import { inBounds } from "../../coreState/utils/Functions";
+import { outlineRect } from "../Pipeline";
+import { FILLED_COLOR } from "../Theme";
 
 export const drawCursor = (canvas, board, controller, windowSize, xCellSize, yCellSize) => {
     var [x, y] = controller.gridCursor(
@@ -17,4 +17,8 @@ export const drawCursor = (canvas, board, controller, windowSize, xCellSize, yCe
       FILLED_COLOR.getHex()
     );
     
+}
+
+export const updateCursor = (cursor) => {
+  
 }
