+++ src/components/game/graphics/objects/Piece.js
@@ -26,6 +26,16 @@ export const drawPiece = (canvas, piece, x0, y0, xCellSize, yCellSize) => {
                 );
             }
         }
-        
+    }
+}
+
+
+export const updatePiece = (piece) => {
+    // Fill in cells from the coreState current piece.
+    var [x, y] = [0, 0];
+    // TODO: Improve
+    if (piece) {
+        var mainCell = piece.mainCell;
+        mainCell.idleUpdate();
     }
 }
