+++ src/components/game/graphics/objects/Targets.js
@@ -1,4 +1,4 @@
-import { drawRect, outlineRect } from "../../Pipeline";
+import { drawRect, outlineRect } from "../Pipeline";
 
 export const drawTargets = (canvas, targets, stage, xCellSize, yCellSize) => {
     // Draw outlines of Targets
